Added abandoning audio focus when exiting PlayListView.java.
Implemented the automatic ducking of volume when another app requests access to audio focus i.e. if the gps wants to say something while the music is playing, the volume of the music will be lowered temporarily then raised again when the gps loses audio focus.

Music player should pause when the phone is called but needs to be tested